.claude/hooks/requirements.txt
openai>=1.0.0
elevenlabs>=0.2.0
python-dotenv>=1.0.0
